refactor inventory and add equipment view and template
