+++ inventory/serializers.py
@@ -1,6 +1,6 @@
 from rest_framework import serializers
 
-from .models import Item
+from inventory.models.item import Item
 
 
 class ItemSerializer(serializers.ModelSerializer):
