+++ inventory/templates/inventory/equipment/all_equipment.html
@@ -1,0 +1,35 @@
+{% extends 'website/base.html' %}
+{% load static %}
+
+{% block head %}
+
+{% endblock %}
+
+{% block content %}
+
+    <div class="row">
+        <div class="col s8">
+            <h2>Utstyr hos Hackerspace</h2>
+        </div>
+        <div class="col s4">
+            <p>Her kan du finne informasjon om hvilket utstyr som er tilgjengelig i Hackerspace.</p>
+        </div>
+        <div class="col s12">
+            <ul class="collection">
+                {% for item in items %}
+                    <li class="collection-item avatar valign-wrapper">
+                        <img src="{% if item.thumbnail %}{{ item.thumbnail.thumb_url }}{% else %}{% static 'inventory/img/unknown.png'%}{% endif %}" class="circle">
+                        <h3>{{ item.name }}</h3>
+                        <p>{{ item.description }}</p>
+                        {% if item.inventory_link %}
+                            <a href="{{ item.inventory_link }}">Til lager</a>
+                        {% endif %}
+                    </li>
+                {% empty %}
+                    <h1>Her var det dessverre tomt:(</h1>
+                {% endfor %}
+            </ul>
+        </div>
+    </div>
+
+{% endblock %}
